% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamesPlayed.R
\name{gamesPlayed}
\alias{gamesPlayed}
\title{Boogie Cousins Games Played Function}
\usage{
gamesPlayed(year)
}
\arguments{
\item{year}{For what year do you want to see how many games Boogie played?}
}
\description{
This function allows you to see how many games played DeMarcus Cousins Played Each Year.
}
\examples{
gamesPlayed(2011)
}
\keyword{nba}
