% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldestPlayer.R
\name{oldestPlayer}
\alias{oldestPlayer}
\title{Oldest Player Per NBA Season}
\usage{
oldestPlayer(year)
}
\arguments{
\item{year}{For what year do you want to see the oldest player?}
}
\description{
This function prints the oldest player in an NBA season, with the player with the most games played breaking ties.
}
\examples{
oldestPlayer(2011)
}
\keyword{nba}
