% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyCorr.R
\name{yearlyCorr}
\alias{yearlyCorr}
\title{NBA Season Stats Correlation Matrix}
\usage{
yearlyCorr(year)
}
\arguments{
\item{year}{For what year do you want to calculate a correlation matrix for?}
}
\description{
This function calculates a cor plot of NBA season statistics for a specific year.
}
\examples{
yearlyCorr(2011)
}
\keyword{nba}
